tf_image_requirements.txt
Cython
numpy
pandas
scikit-learn
Pillow
tfa-nightly
tensorflow-datasets
matplotlib
git+https://github.com/albu/albumentations
